w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/forms/the-input-element/datetime-expected.txt

Date & Time Inputs


FAIL date type support on input element assert_equals: expected "date" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:27:35
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:27:9)
PASS [date] The value must be a valid global date and time string 
PASS [date] The min attribute must have a value that is a valid global date and time string 
PASS [date] The max attribute must have a value that is a valid global date and time string 
PASS [date] The step attribute must be expressed in seconds 
PASS [date] stepUp method support on input 'date' element 
PASS [date] stepDown method support on input 'date' element 
FAIL [time] time type support on input element assert_equals: expected "time" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:42:35
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:42:9)
PASS [time] The value must be a valid global date and time string 
PASS [time] The min attribute must have a value that is a valid global date and time string 
PASS [time] The max attribute must have a value that is a valid global date and time string 
PASS [time] The step attribute must be expressed in seconds 
PASS [time] stepUp method support on input 'time' element 
PASS [time] stepDown method support on input 'time' element 
FAIL datetime type support on input element assert_equals: expected "datetime" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:57:35
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:57:9)
PASS [datetime] The must be a valid global date and time string 
PASS [datetime] The min attribute must have a value that is a valid global date and time string 
PASS [datetime] The max attribute must have a value that is a valid global date and time string 
PASS [datetime] The step attribute must be expressed in seconds 
PASS [datetime] stepUp method support on input 'datetime' element 
PASS [datetime] stepDown method support on input 'datetime' element 
FAIL month type support on input element assert_equals: expected "month" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:72:35
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:72:9)
PASS [month] The value must be a value that is a valid global date and time string 
PASS [month] The min attribute must have a value that is a valid global date and time string 
PASS [month] The max attribute must have a value that is a valid global date and time string 
PASS [month] The step attribute must be expressed in seconds 
PASS [month] stepUp method support on input 'month' element 
PASS [month] stepDown method support on input 'month' element 
FAIL week type support on input element assert_equals: expected "week" but got "text"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:87:35
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/datetime.html:87:9)
PASS [week] The value must be a value that is a valid global date and time string 
PASS [week] The min attribute must have a value that is a valid global date and time string 
PASS [week] The max attribute must have a value that is a valid global date and time string 
PASS [week] The step attribute must be expressed in seconds 
PASS [week] stepUp method support on input 'week' element 
PASS [week] stepDown method support on input 'week' element 
